pyproject.toml
[project]
name = "polls-service"
version = "0.1.0"
description = ""
authors = [
    { name = "Kirill67tyar", email = "kirillbogomolov.ric@gmail.com" },
]
requires-python = ">=3.12,<3.13"
dynamic = ["dependencies"]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
django = "^6.0.4"
djangorestframework = "^3.17.1"
psycopg = { extras = ["binary"], version = "^3.2" }
python-dotenv = "^1.0.1"
ipython = "^9.11.0"
django-extensions = "^4.1"
django-debug-toolbar = "^6.3.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.0"
pre-commit = "^4.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"config/settings.py" = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
exclude = ["config/settings.py"]
